Move upload observation code into ObsEditProvider (for single uploads)
