+++ src/realmModels/Observation.js
@@ -1,10 +1,6 @@
 import { Realm } from "@realm/react";
-// eslint-disable-next-line import/no-cycle
-import { createObservation, createOrUpdateEvidence, updateObservation } from "api/observations";
-import inatjs from "inaturalistjs";
 import uuid from "react-native-uuid";
 import { createObservedOnStringForUpload } from "sharedHelpers/dateAndTime";
-import emitUploadProgress from "sharedHelpers/emitEvent";
 
 import Application from "./Application";
 import Comment from "./Comment";
@@ -232,156 +228,6 @@ class Observation extends Realm.Object {
     const obsList = Observation.filterUnsyncedObservations( realm );
     const unsyncedObs = obsList.filtered( `uuid == "${obs.uuid}"` );
     return unsyncedObs.length > 0;
-  };
-
-  static markRecordUploaded = ( recordUUID, type, response, realm ) => {
-    const { id } = response.results[0];
-
-    const record = realm.objectForPrimaryKey( type, recordUUID );
-    realm?.write( ( ) => {
-      record.id = id;
-      record._synced_at = new Date( );
-    } );
-  };
-
-  static uploadToServer = async (
-    evidenceUUID: string,
-    type: string,
-    params: Object,
-    apiEndpoint: Function,
-    realm: any,
-    options: Object,
-    observationUUID: string
-  ) => {
-    const response = await createOrUpdateEvidence(
-      apiEndpoint,
-      observationUUID,
-      params,
-      options
-    );
-    if ( response ) {
-      Observation.markRecordUploaded( evidenceUUID, type, response, realm );
-    }
-  };
-
-  static uploadEvidence = async (
-    evidence: Array<Object>,
-    type: string,
-    apiSchemaMapper: Function,
-    observationId: number,
-    apiEndpoint: Function,
-    realm: any,
-    options: Object,
-    observationUUID: string,
-    forceUpload: boolean
-  ): Promise<any> => {
-    // only try to upload evidence which is not yet on the server
-    const unsyncedEvidence = forceUpload
-      ? evidence
-      : evidence.filter( item => !item.wasSynced( ) );
-
-    const responses = await Promise.all( unsyncedEvidence.map( item => {
-      const currentEvidence = item.toJSON( );
-      const evidenceUUID = currentEvidence.uuid;
-
-      // Remove all null values, b/c the API doesn't seem to like them
-      const newPhoto = {};
-      const photo = currentEvidence?.photo;
-      Object.keys( photo ).forEach( k => {
-        if ( photo[k] !== null ) {
-          newPhoto[k] = photo[k];
-        }
-      } );
-
-      currentEvidence.photo = newPhoto;
-
-      const params = apiSchemaMapper( observationId, currentEvidence );
-      return Observation.uploadToServer(
-        evidenceUUID,
-        type,
-        params,
-        apiEndpoint,
-        realm,
-        options,
-        observationUUID
-      );
-    } ) );
-    // eslint-disable-next-line consistent-return
-    return responses[0];
-  };
-
-  static uploadObservation = async ( obs, apiToken, realm ) => {
-    // every observation and observation photo counts for a total of 1 progress
-    // we're showing progress in 0.5 increments: when an upload of obs/obsPhoto starts
-    // and when the upload of obs/obsPhoto successfully completes
-    emitUploadProgress( obs.uuid, 0.5 );
-    const obsToUpload = Observation.mapObservationForUpload( obs );
-    const options = { api_token: apiToken };
-
-    // Remove all null values, b/c the API doesn't seem to like them for some
-    // reason (might be an error with the API as of 20220801)
-    const newObs = {};
-    Object.keys( obsToUpload ).forEach( k => {
-      if ( obsToUpload[k] !== null ) {
-        newObs[k] = obsToUpload[k];
-      }
-    } );
-
-    let response;
-
-    // First upload the photos/sounds (before uploading the observation itself)
-    const hasPhotos = obs?.observationPhotos?.length > 0;
-
-    await Promise.all( [
-      hasPhotos
-        ? Observation.uploadEvidence(
-          obs.observationPhotos,
-          "ObservationPhoto",
-          ObservationPhoto.mapPhotoForUpload,
-          null,
-          inatjs.photos.create,
-          realm,
-          options
-        )
-        : null
-    ] );
-
-    const wasPreviouslySynced = obs.wasSynced( );
-    const uploadParams = {
-      observation: { ...newObs },
-      fields: { id: true }
-    };
-
-    if ( wasPreviouslySynced ) {
-      response = await updateObservation( {
-        ...uploadParams,
-        id: newObs.uuid,
-        ignore_photos: true
-      }, options );
-    } else {
-      response = await createObservation( uploadParams, options );
-    }
-
-    const { uuid: obsUUID } = response.results[0];
-
-    await Promise.all( [
-      Observation.markRecordUploaded( obs.uuid, "Observation", response, realm ),
-      // Next, attach the uploaded photos/sounds to the uploaded observation
-      hasPhotos
-        ? Observation.uploadEvidence(
-          obs.observationPhotos,
-          "ObservationPhoto",
-          ObservationPhoto.mapPhotoForAttachingToObs,
-          obsUUID,
-          inatjs.observation_photos.create,
-          realm,
-          options,
-          obsUUID,
-          true
-        )
-        : null
-    ] );
-    return response;
   };
 
   static schema = {
