+++ src/sharedHelpers/emitUploadProgress.js
