+++ src/sharedHooks/useUploadObservations.js
@@ -4,21 +4,19 @@ import {
   activateKeepAwake,
   deactivateKeepAwake
 } from "@sayem314/react-native-keep-awake";
-import { RealmContext } from "providers/contexts";
-import { useEffect, useState } from "react";
-import Observation from "realmModels/Observation";
+import { ObsEditContext } from "providers/contexts";
+import { useContext, useEffect, useState } from "react";
 import useApiToken from "sharedHooks/useApiToken";
 
-const { useRealm } = RealmContext;
-
 const useUploadObservations = ( allObsToUpload: Array<Object> ): Object => {
+  const obsEditContext = useContext( ObsEditContext );
+  const uploadObservation = obsEditContext?.uploadObservation;
   const [uploadInProgress, setUploadInProgress] = useState( false );
   const [shouldUpload, setShouldUpload] = useState( false );
   const [currentUploadIndex, setCurrentUploadIndex] = useState( 0 );
   const [error, setError] = useState( null );
   const [uploads, setUploads] = useState( [] );
   const [uploadedUUIDs, setUploadedUUIDS] = useState( [] );
-  const realm = useRealm( );
   const apiToken = useApiToken( );
 
   const cleanup = ( ) => {
@@ -36,11 +34,7 @@ const useUploadObservations = ( allObsToUpload: Array<Object> ): Object => {
         observationToUpload.uuid
       ] );
       try {
-        await Observation.uploadObservation(
-          observationToUpload,
-          apiToken,
-          realm
-        );
+        await uploadObservation( observationToUpload );
       } catch ( e ) {
         console.warn( e );
         setError( e.message );
@@ -70,8 +64,8 @@ const useUploadObservations = ( allObsToUpload: Array<Object> ): Object => {
     apiToken,
     shouldUpload,
     currentUploadIndex,
-    realm,
-    uploads
+    uploads,
+    uploadObservation
   ] );
 
   // // Fake upload in progress
